requirements.txt
# API Clients
praw==7.7.1          # Reddit API wrapper
tweepy==4.14.0       # Twitter API client (optional)

# Data Processing
pandas==2.1.1
numpy==1.26.0
python-dotenv==1.0.0

# NLP & Sentiment Analysis
nltk==3.8.1
vaderSentiment==3.3.2
spacy==3.7.2

# Visualization
plotly==5.18.0
wordcloud==1.9.2

# Web Application
streamlit==1.28.0

# Testing
pytest==7.4.2